Removed DNX/.NET core stuff, refactored program/main out
