+++ RawBencher.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
-VisualStudioVersion = 14.0.23107.0
+VisualStudioVersion = 14.0.25123.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RawBencher", "RawBencher\RawBencher.csproj", "{A4AA317A-6958-4D0C-B2EE-026005FD2C2E}"
 EndProject
@@ -16,8 +16,6 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EF6.Bencher.Persistence", "EF6\Persistence\EF6.Bencher.Persistence.csproj", "{AC909E22-26D0-47CF-98BC-DD07309D3A77}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OakDynamicDb.Bencher", "OakDynamicDb.Bencher\OakDynamicDb.Bencher.csproj", "{106CED47-76BC-477E-9FFF-3E6042F874EB}"
-EndProject
-Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "RawBencher.DNX", "RawBencher.DNX\RawBencher.DNX.xproj", "{C5FD24A4-9E79-426D-BD58-754CF2403CD5}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EF7.Bencher.Model", "EF7.Bencher.Model\EF7.Bencher.Model.csproj", "{6E36AEC5-29FE-4B53-9C46-AB6FC7075094}"
 EndProject
@@ -112,18 +110,6 @@ Global
 		{106CED47-76BC-477E-9FFF-3E6042F874EB}.Release|Any CPU.Build.0 = Release|Any CPU
 		{106CED47-76BC-477E-9FFF-3E6042F874EB}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
 		{106CED47-76BC-477E-9FFF-3E6042F874EB}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{C5FD24A4-9E79-426D-BD58-754CF2403CD5}.Debug|.NET.ActiveCfg = Debug|Any CPU
-		{C5FD24A4-9E79-426D-BD58-754CF2403CD5}.Debug|.NET.Build.0 = Debug|Any CPU
-		{C5FD24A4-9E79-426D-BD58-754CF2403CD5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C5FD24A4-9E79-426D-BD58-754CF2403CD5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C5FD24A4-9E79-426D-BD58-754CF2403CD5}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{C5FD24A4-9E79-426D-BD58-754CF2403CD5}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{C5FD24A4-9E79-426D-BD58-754CF2403CD5}.Release|.NET.ActiveCfg = Release|Any CPU
-		{C5FD24A4-9E79-426D-BD58-754CF2403CD5}.Release|.NET.Build.0 = Release|Any CPU
-		{C5FD24A4-9E79-426D-BD58-754CF2403CD5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C5FD24A4-9E79-426D-BD58-754CF2403CD5}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C5FD24A4-9E79-426D-BD58-754CF2403CD5}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{C5FD24A4-9E79-426D-BD58-754CF2403CD5}.Release|Mixed Platforms.Build.0 = Release|Any CPU
 		{6E36AEC5-29FE-4B53-9C46-AB6FC7075094}.Debug|.NET.ActiveCfg = Debug|Any CPU
 		{6E36AEC5-29FE-4B53-9C46-AB6FC7075094}.Debug|.NET.Build.0 = Debug|Any CPU
 		{6E36AEC5-29FE-4B53-9C46-AB6FC7075094}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
