+++ RawBencher/Benchers/BencherBase.cs
@@ -290,7 +290,7 @@ namespace RawBencher.Benchers
 		{
 			string assemblyVersion;
 			string fileVersion;
-			GetVersionStrings(Program.GetAssembly(t), out fileVersion, out assemblyVersion);
+			GetVersionStrings(BencherUtils.GetAssembly(t), out fileVersion, out assemblyVersion);
 			return string.Format(formatString, assemblyVersion, fileVersion);
 		}
 
@@ -374,9 +374,7 @@ namespace RawBencher.Benchers
 			if(toProbe != null)
 			{
 				assemblyVersion = toProbe.GetName().Version.ToString();
-#if !DNXCORE50 // Location not available
 				fileVersion = FileVersionInfo.GetVersionInfo(toProbe.Location).FileVersion;
-#endif
 			}
 		}
