+++ RawBencher/Benchers/DataTableBencher.cs
@@ -1,5 +1,4 @@
-﻿#if !DNXCORE50
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Data;
 using System.Data.SqlClient;
@@ -91,4 +90,3 @@ namespace RawBencher.Benchers
 		#endregion
 	}
 }
-#endif
