+++ RawBencher/Benchers/EntityFramework7NoChangeTrackingBencher.cs
@@ -1,5 +1,4 @@
-﻿#if !(DNXCORE50 || DNX451)
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -68,4 +67,3 @@ namespace RawBencher.Benchers
 		#endregion
 	}
 }
-#endif
