+++ RawBencher/Benchers/HandCodedBencherUsingBoxing.cs
@@ -133,9 +133,7 @@ namespace RawBencher.Benchers
 					soh.ModifiedDate = (DateTime)values[25];
 					headers.Add(soh);
 				}
-#if !DNXCORE50
 				reader.Close();
-#endif
 				reader.Dispose();
 				con.Close();
 			}
