+++ RawBencher/Benchers/LLBLGenProResultsetCachingBencher.cs
@@ -1,5 +1,4 @@
-﻿#if !(DNXCORE50 || DNX451)
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -77,4 +76,3 @@ namespace RawBencher.Benchers
 		}
 	}
 }
-#endif
