+++ RawBencher/Benchers/LinqToSqlNormalBencher.cs
@@ -1,5 +1,4 @@
-﻿#if !(DNXCORE50 || DNX451)
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Data.Linq;
 using System.Linq;
@@ -109,4 +108,3 @@ namespace RawBencher.Benchers
 		}
 	}
 }
-#endif
