+++ RawBencher/Benchers/OakDynamicDbDtoBencher.cs
@@ -1,5 +1,4 @@
-﻿#if !(DNXCORE50 || DNX451)
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Data.SqlClient;
 using System.Linq;
@@ -68,4 +67,3 @@ namespace RawBencher.Benchers
 	}
 }
 
-#endif
