+++ RawBencher/Massive.cs
@@ -1,4 +1,3 @@
-#if !DNXCORE50
 using System;
 using System.Collections.Generic;
 using System.Collections.Specialized;
@@ -887,4 +886,3 @@ namespace VanillaMassive {
 		}
 	}
 }
-#endif
