+++ RawBencher/RawBencher.csproj
@@ -205,9 +205,11 @@
     <Compile Include="Benchers\LLBLGenProResultsetCachingBencher.cs" />
     <Compile Include="Benchers\LLBLGenProNoChangeTrackingBencher.cs" />
     <Compile Include="Benchers\LLBLGenProNormalBencher.cs" />
+    <Compile Include="BencherUtils.cs" />
     <Compile Include="BenchResult.cs" />
     <Compile Include="Benchers\PetaPocoBencher.cs" />
     <Compile Include="Massive.cs" />
+    <Compile Include="OriginalRunner.cs" />
     <Compile Include="PetaPoco.cs" />
     <Compile Include="Product.cs" />
     <Compile Include="Program.cs" />
